code structure refactoring
