+++ src/MAIN/auto_simulator.py
@@ -3,7 +3,7 @@ import logging
 import time
 import sys
 
-from src.simulator import WolfSimulator
+from src.SIMULATOR.simulator import WolfSimulator
 from src.utils import check_file, check_config
 
 logger = logging.getLogger()
